Move quasar.qscript.analysis.{Cardinality, Cost} to separate files.
